requirements.txt
ipykernel==6.27.1
torch==2.1.1
yapf==0.40.2
numpy==1.26.2
absl-py==2.0.0